address comments and refactor
