+++ protocol/x/sending/keeper/transfer.go
@@ -104,8 +104,11 @@ func (k Keeper) ProcessDepositToSubaccount(
 
 	// Emit gauge metric with labels if deposit to subaccount succeeds.
 	if err == nil {
-		metrics.ContextuallySetGaugeWithLabels(
+		metrics.EmitTelemetryWithLabelsForExecMode(
 			ctx,
+			// sdk.ExecModeFinalize is used here to ensure metrics are only emitted in the Finalize ExecMode.
+			[]sdk.ExecMode{sdk.ExecModeFinalize},
+			telemetry.SetGaugeWithLabels,
 			[]string{
 				types.ModuleName,
 				metrics.ProcessDepositToSubaccount,
@@ -114,8 +117,6 @@ func (k Keeper) ProcessDepositToSubaccount(
 			[]gometrics.Label{
 				metrics.GetLabelForIntValue(metrics.AssetId, int(msgDepositToSubaccount.AssetId)),
 			},
-			// sdk.ExecModeFinalize is used here to ensure metrics are only emitted in the Finalize ExecMode.
-			[]sdk.ExecMode{sdk.ExecModeFinalize},
 		)
 
 		// Add deposit event to Indexer block message.
